Added status updates to CastPlusPlugin
The CastPlusPlugin now includes status updates for various session events. This was achieved by introducing an EventChannel and EventSink, which are used to send status updates as the session progresses. These updates include when a session is starting, has started, resumed, failed to start or resume, ended or suspended. Additionally, error messages are also sent through this channel if the remote media client is null.